pyproject.toml
[project]
name = "crypto-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.2.1",
    "cryptography>=46.0.7",
    "dbt>=1.0.0.40.16",
    "dbt-postgres>=1.10.0",
    "duckdb>=1.5.2",
    "loguru>=0.7.3",
    "pendulum>=3.2.0",
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "rabbitmq>=0.2.0",
    "websocket-client>=1.9.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.10",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N", "UP", "B", "S", "SIM", "RUF", "T20"]
